Switch to cmake
Former-commit-id: af740a29bfb1bfbb709c550c62dd80646cf7b381

+++ Lithium/src/Lithium/Platform/D3D11/D3D11Window.cpp
@@ -93,7 +93,7 @@ namespace Li
 			break;
 		}
 		default:
-			Log::CoreError("Unknown fullscreen type {}.", type);
+			Log::CoreError("Unknown fullscreen type {}.", static_cast<unsigned int>(type));
 			break;
 		}
 	}
